+++ sourced/ml/utils/__init__.py
@@ -1,4 +1,4 @@
 from sourced.ml.utils.bigartm import install_bigartm
-from sourced.ml.utils.engine import create_engine, create_spark, add_engine_args, add_spark_args, \
-    EngineConstants
+from sourced.ml.utils.spark import spark_config, create_spark, add_spark_args, SparkDefault
+from sourced.ml.utils.engine import create_engine, add_engine_args, EngineConstants
 from sourced.ml.utils.pickleable_logger import PickleableLogger
